backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.18
docker==7.1.0
boto3==1.35.76
paramiko==3.5.0
apscheduler==3.10.4
aiofiles==24.1.0
httpx==0.28.1
